move run_me functions into skellytracker folder to be accessible to __main__
